cambiado tipo devuelto en método size de testlineasfacturadao a long

+++ src/test/java/es/cipfpbatoi/TestLineaFacturaDAO.java
@@ -192,7 +192,7 @@ class TestLineaFacturaDAO {
 	@Test
 	@Order(1)
 	void testSize() {
-		int respuestaObtenida = capaDao.size();
+		long respuestaObtenida = capaDao.size();
 		assertEquals(-1, respuestaObtenida);
 	}
